+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [1.4.0] - 2025-11-26
+### Added
+- added the option to choose which product to use for parcel locker and courier methods in the module settings
+
+### Changed
+- changed so that API logins are not divided into products, but the same ones are used for all
+
+### Updated
+- itella-api to 2.5.0
+
 ## [1.3.4] - 2025-11-10
 ### Updated
 - itella-api to 2.4.10